requirements.txt
pyserial
requests
pytz
